Remove unnecessary casts to floats

+++ software/o_c_REV/streams_lorenz_generator.h
@@ -60,8 +60,10 @@ class LorenzGenerator {
   }
 
   inline void set_rho2(int16_t rho) {
-    rho2_ = ((double)(rho) * (1 << 13)) + 24.0 * (1 << 24) ; // was 12
-    c2_ = (double)(rho + (6 << 3)) * (1 << 13) ; // was 13
+    // rho2_ = ((double)(rho) * (1 << 13)) + 24.0 * (1 << 24) ; // was 12
+    // c2_ = (double)(rho + (6 << 3)) * (1 << 13) ; // was 13
+    rho2_ = (rho * (1 << 13)) + 24.0 * (1 << 24) ; // was 12
+    c2_ = (rho + (6 << 3)) * (1 << 13) ; // was 13
   }
 
   inline void set_out_a(uint8_t out_a) {
